requirements.txt
numpy==1.22.4
pandas 
scikit-learn
pytesseract
geonamescache
spacy
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.0.0/en_core_web_lg-3.0.0-py3-none-any.whl
tensorflow
keras
pdf2image
opencv-python
pandas
sklearn
tensorflow-cpu  
markupsafe==1.1.1
Pillow
azure.storage.blob
h5py
openpyxl